requirements-docker.txt
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
google-api-core==2.10.0
google-auth==2.11.0
google-cloud-documentai==2.0.0
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proto-plus==1.22.1
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2